pyproject.toml
[project]
name = "fastapi-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.3.2",
    "cryptography>=46.0.3",
    "fastapi>=0.124.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.45",
    "uvicorn>=0.38.0",
]
